pyproject.toml
[tool.poetry]
name = "bicephalus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.4"
opentelemetry-api = "^1.16.0"
opentelemetry-sdk = "^1.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
